adjust edit-movie

+++ admin/pages/edit-movie-sql.php
@@ -19,22 +19,10 @@
 	$objResult = mysqli_fetch_array($objQuery,MYSQLI_ASSOC);
 
     $movie_id = $_POST["movie_id"];
-    $movie_name = mysqli_real_escape_string($con,$_POST["movie_name"]);
-    $movie_description = mysqli_real_escape_string($con,$_POST["movie_description"]);
-    $movie_trailer = mysqli_real_escape_string($con,$_POST["movie_trailer"]);
-    $director_info = mysqli_real_escape_string($con,$_POST["director_info"]);
-    $movie_length = $_POST["movie_length"];
-    $r_Date = $_POST["releaseDate"];
-    $sp_date = $_POST["start_promote"];
-    $np_date = $_POST["end_promote"];
-    // $releaseDate = date ('Y-m-d H:i:s', strtotime($r_Date));
-
-    $releaseDate = date ('Y-m-d', strtotime($r_Date));
-    $start_promote = date ('Y-m-d', strtotime($sp_date));
-    $end_promote = date ('Y-m-d', strtotime($np_date));
 
     if(!empty($_POST["movie_name"]))
     {
+      $movie_name = mysqli_real_escape_string($con,$_POST["movie_name"]);
         $sql = "UPDATE movieinfo SET movie_name = '$movie_name' WHERE movie_id =  $movie_id"; 
         if (!mysqli_query($con, $sql)) {
 			die('Error: ' . mysqli_error($con));
@@ -44,6 +32,7 @@
     }
     if(!empty($_POST["movie_description"]))
     {
+      $movie_description = mysqli_real_escape_string($con,$_POST["movie_description"]);
         $sql = "UPDATE movieinfo SET movie_description = '$movie_description' WHERE movie_id =  $movie_id"; 
          if (!mysqli_query($con, $sql)) {
 			die('Error: ' . mysqli_error($con));
@@ -52,7 +41,7 @@
     }
     if(!empty($_POST["movie_trailer"]))
     {
-        
+      $movie_trailer = mysqli_real_escape_string($con,$_POST["movie_trailer"]);
       $sql = "UPDATE movieinfo SET movie_trailer = '$movie_trailer' WHERE movie_id =  $movie_id"; 
       if (!mysqli_query($con, $sql)) {
 			die('Error: ' . mysqli_error($con));
@@ -61,6 +50,7 @@
     }
     if(!empty($_POST["director_info"]))
     {
+      $director_info = mysqli_real_escape_string($con,$_POST["director_info"]);
       $sql = "UPDATE movieinfo SET director_info = '$director_info' WHERE movie_id =  $movie_id"; 
       if (!mysqli_query($con, $sql)) {
 			die('Error: ' . mysqli_error($con));
@@ -69,6 +59,7 @@
     }
     if(!empty($_POST["movie_length"]))
     {
+      $movie_length = $_POST["movie_length"];
       $sql = "UPDATE movieinfo SET movie_length = $movie_length WHERE movie_id =  $movie_id";
       if (!mysqli_query($con, $sql)) {
 			die('Error: ' . mysqli_error($con));
@@ -77,6 +68,8 @@
     }
     if(!empty($_POST["releaseDate"]))
     {
+      $r_Date = $_POST["releaseDate"];
+      $releaseDate = date ('Y-m-d', strtotime($r_Date));
       $sql = "UPDATE movieinfo SET releaseDate = '$releaseDate' WHERE movie_id =  $movie_id";
         if (!mysqli_query($con, $sql)) {
 			die('Error: ' . mysqli_error($con));
@@ -85,6 +78,8 @@
     }
     if(!empty($_POST["start_promote"]))
     {
+      $sp_date = $_POST["start_promote"];
+      $start_promote = date ('Y-m-d', strtotime($sp_date));
       $sql = "UPDATE movieinfo SET start_promote = '$start_promote' WHERE movie_id =  $movie_id";
       if (!mysqli_query($con, $sql)) {
 			die('Error: ' . mysqli_error($con));
@@ -93,6 +88,8 @@
     }
     if(!empty($_POST["end_promote"]))
     {
+      $np_date = $_POST["end_promote"];
+      $end_promote = date ('Y-m-d', strtotime($np_date));
       $sql = "UPDATE movieinfo SET end_promote = '$end_promote' WHERE movie_id =  $movie_id";
       if (!mysqli_query($con, $sql)) {
 			die('Error: ' . mysqli_error($con));
@@ -112,20 +109,31 @@
       $id = $a["movie_id"];
     }
 
-    $filename = "../../home/img/poster/".$id.".jpg";
-    unlink($filename);
 
-    $filename = "../../home/img/promote/".$id.".jpg";
-    unlink($filename);
-    
-    $filename = $_FILES["inputPosterImage"]["movie_name"];
-    $path = "../../home/img/poster/".$id.".jpg";
-    move_uploaded_file($_FILES['inputPosterImage']['tmp_name'],$path);
+    if($_FILES['inputPosterImage']['movie_name'] == '')
+    {
+      $filename = "../../home/img/poster/".$id.".jpg";
+      unlink($filename);
 
+      $filename = $_FILES["inputPosterImage"]["movie_name"];
+      $path = "../../home/img/poster/".$id.".jpg";
+      move_uploaded_file($_FILES['inputPosterImage']['tmp_name'],$path);
+      echo "<script> alert('Edit inputPosterImage Information succesful!'); </script>";
+  
+    }
+
+    if($_FILES['inputPosterImage']['movie_name'] == '')
+    {
+      $filename = "../../home/img/promote/".$id.".jpg";
+      unlink($filename);    
+
+      $filename = $_FILES["inputPromoteImage"]["movie_name"];
+      $path = "../../home/img/promote/".$id.".jpg";
+      move_uploaded_file($_FILES['inputPromoteImage']['tmp_name'],$path);
+      echo "<script> alert('Edit inputPosterImage Information succesful!'); </script>";
+  
+    }
     
-    $filename = $_FILES["inputPromoteImage"]["movie_name"];
-    $path = "../../home/img/promote/".$id.".jpg";
-    move_uploaded_file($_FILES['inputPromoteImage']['tmp_name'],$path);
 
     echo "<script> alert('Edit Movie Information succesful!'); </script>";
